feat: implement SettingsDrawer and associated settings panels; refactor LibraryPage and related components
