+++ src/app/AppRoutes.tsx
@@ -7,7 +7,7 @@ import { lazy } from "react";
 import { BrowserRouter, Route, Routes } from "react-router";
 
 const BoardPage = lazy(() => import("@pages/BoardPage"));
-const LibraryPage = lazy(() => import("@pages/LibraryPage/LibraryPage"));
+const LibraryPage = lazy(() => import("@pages/library/LibraryPage"));
 const AboutPage = lazy(() => import("@pages/AboutPage"));
 
 export function AppRoutes() {
