+++ src/app/drawers/settings/panels/AISettings.tsx
